Comment out failing import in test context
The idea behind importing the contents of the project's package into
context.py is that it makes your modules available to your tests.

Importing them all automatically triggered a linting error, and it's not
how I actually do it anyway. I import each new of the package's modules
specifically. So I'm documenting that here.